Disable GCP vm deletion on production until stable release

+++ prog/gcp_vm/nexus.rb
@@ -224,10 +224,13 @@ class Prog::GcpVm::Nexus < Prog::Base
   label def destroy
     DB.transaction do
       gcp_vm.update(display_state: "deleting")
-      gcp_client = Hosting::GcpApis.new
-      gcp_client.delete_vm(gcp_vm.name, "#{gcp_vm.location}-a")
-      if gcp_vm.has_static_ipv4
-        gcp_client.release_ipv4(gcp_vm.name, gcp_vm.location)
+      # TODO:: disable vm deletion until stable release
+      if !Config.production?
+        gcp_client = Hosting::GcpApis.new
+        gcp_client.delete_vm(gcp_vm.name, "#{gcp_vm.location}-a")
+        if gcp_vm.has_static_ipv4
+          gcp_client.release_ipv4(gcp_vm.name, gcp_vm.location)
+        end
       end
       strand.children.each { _1.destroy }
       gcp_vm.projects.map { gcp_vm.dissociate_with_project(_1) }
